asset_discovery/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scapy==2.5.0
python-nmap==0.7.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.2.1
apscheduler==3.10.4
netifaces-binary==0.11.0
requests==2.32.3
httpx==0.27.0
loguru==0.7.2
asyncpg==0.29.0
reportlab==4.0.0
python-docx==0.8.12
